second server sending pages and recieving password from recording client
