+++ public/main.js
@@ -1,4 +1,4 @@
-  var client = new WebSocket( 'ws://192.168.0.16:8084/');
+  var client = new WebSocket( 'ws://10.0.0.166:8084/');
   let socket= io(client)
 
     var canvas = document.getElementById('videoCanvas');
